moved company model to users app
