+++ pplatform/catalog/models.py
@@ -7,25 +7,7 @@ from django.utils import timezone
 from django.utils.text import slugify
 
 # from django.contrib.auth import get_user_model
-# from pplatform.users.models import CustomUser
-
-
-class Company(models.Model):
-    name = models.CharField(max_length=200, db_index=True)
-    slug = models.SlugField(max_length=200, unique=True)
-    # owner = models.ForeignKey(CustomUser, related_name='companies_own', on_delete=models.CASCADE)
-    created = models.DateTimeField(auto_now_add=True)
-
-    class Meta:
-        ordering = ("name",)
-        verbose_name = "company"
-        verbose_name_plural = "companies"
-
-    def get_absolute_url(self):
-        return reverse("catalog:product_list_by_company", args=[self.slug])
-
-    def __str__(self):
-        return self.name
+from pplatform.users.models import Company, CustomUser
 
 
 class Category(models.Model):
@@ -61,6 +43,13 @@ class Product(models.Model):
     )
     name = models.CharField(max_length=200, db_index=True)
     slug = models.SlugField(max_length=200, db_index=True)
+    created_by = models.ForeignKey(
+        CustomUser,
+        on_delete=models.CASCADE,
+        related_name="products_created",
+        blank=True,
+        null=True,
+    )
     image = models.ImageField(upload_to="products/%Y/%m/%d", blank=True)
     description = models.TextField(blank=True)
     # price = models.DecimalField(max_digits=10, decimal_places=2)
